created household page integration, refactored chore display, implemented routing
