+++ package.json
@@ -19,7 +19,8 @@
     "firebase": "^11.6.0",
     "react": "^19.0.0",
     "react-datepicker": "^8.3.0",
-    "react-dom": "^19.0.0"
+    "react-dom": "^19.0.0",
+    "react-router-dom": "^7.5.1"
   },
   "devDependencies": {
     "@testing-library/react": "^16.2.0",
