--- types/status.tsx
@@ -1,1 +1,0 @@
-export type Status = 'all' | 'completed' | 'incomplete';
